requirements.txt
pandas>=2.0.0
pdfplumber>=0.10.0,<0.11
matplotlib>=3.7.0
cryptography<42

# Phase 3: Embeddings and Vector Retrieval
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
torch>=2.0.0
tqdm>=4.65.0
datasets>=2.14.0
transformers>=4.30.0
huggingface-hub>=0.34.0,<1.0

# Phase 4: Graph-Constrained RAG and UI
neo4j>=5.0.0
requests>=2.31.0
langgraph>=0.2.0
streamlit>=1.30.0
python-dotenv>=1.0.0
